Replaced page prop and turning mechanism
