+++ src/components/book-components/Page.jsx
@@ -8,10 +8,52 @@ const Page = (props) => {
     height: 100%;
     width: 100%;
 
-    pointer-events: none;
+    pointer-events: ${props.visible ? "all" : "none"};
   `;
 
-  return <Container>{props.children}</Container>;
+  function leftPageClicked() {
+    console.log("Left page clicked");
+
+    const currentPage = props.pageNumber;
+
+    const curPageDiv = Math.floor((currentPage - 1) / 2) + 1;
+
+    if (curPageDiv > 1) {
+      if (currentPage % 2 == 0) {
+        props.updatePage(currentPage - 2);
+      } else {
+        props.updatePage(currentPage - 1);
+      }
+    }
+  }
+
+  function rightPageClicked() {
+    console.log("Right page clicked");
+
+    const currentPage = props.pageNumber;
+
+    const maxPageDiv = Math.floor((maxPage - 1) / 2) + 1;
+
+    const curPageDiv = Math.floor((currentPage - 1) / 2) + 1;
+
+    if (curPageDiv < maxPageDiv) {
+      if (currentPage % 2 == 0) {
+        props.updatePage(currentPage + 1);
+      } else {
+        props.updatePage(currentPage + 2);
+      }
+    }
+  }
+
+  function divClicked() {
+    if (props.pageNumber % 2 == 1) {
+      leftPageClicked();
+    } else {
+      rightPageClicked();
+    }
+  }
+
+  return <Container onClick={divClicked}>{props.children}</Container>;
 };
 
 export default Page;
